+++ PBL A&P 2/src/model/JogadorDao.java
@@ -3,7 +3,7 @@ package model;
 import java.util.List;
 
 public interface JogadorDao extends PessoaDAO{
-	abstract void inserirJogador(Selecao selecao, String nome, int idade, float altura, boolean titular, String nacio, int posicao,int camisa);
+	abstract void inserirJogador(Selecao selecao, String nome, int idade, double altura, boolean titular, String nacio, int posicao,int camisa);
 	abstract void editarJogador(Selecao selecao, int id, int num, String dado);
 	abstract void excluirJogador(Selecao selecao, int id);
 	abstract List<Jogador> listarJogador(Selecao selecao);
